ISSUE #5264 - fix: clear values when switching to fieldless operator
