+++ frontend/src/v5/ui/components/viewer/cards/cardFilters/filterForm/filterForm.helpers.ts
@@ -17,7 +17,7 @@
 
 import { CardFilterOperator } from '../cardFilters.types';
 
-export const getOperatorMaxSupportedValues = (operator: CardFilterOperator) => {
+export const getOperatorMaxFieldsAllowed = (operator: CardFilterOperator) => {
 	switch (operator) {
 		case 'ex':
 		case 'nex':
