+++ frontend/src/v5/ui/components/viewer/cards/cardFilters/filterForm/filterFormValues/filterFormValues.componen.tsx
@@ -15,26 +15,58 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { useFormContext } from 'react-hook-form';
-import { getOperatorMaxSupportedValues } from '../filterForm.helpers';
+import { useFieldArray, useFormContext } from 'react-hook-form';
+import { getOperatorMaxFieldsAllowed } from '../filterForm.helpers';
 import { isTextType } from '../../cardFilters.helpers';
-import { FilterFormTextValues } from './types/filterFormTextValues.component';
 import { FormTextField } from '@controls/inputs/formInputs.component';
+import { ArrayFieldContainer } from '@controls/inputs/arrayFieldContainer/arrayFieldContainer.component';
+import { useEffect } from 'react';
+import { range } from 'lodash';
 
 export const FilterFormValues = ({ type }) => {
-	const { watch } = useFormContext();
-	const formOperator = watch('operator');
-	const valuesInputsCount = Math.min(getOperatorMaxSupportedValues(formOperator), 3);
+	const { control, watch, formState: { errors } } = useFormContext();
+	const { fields, append, remove } = useFieldArray({
+		control,
+		name: 'values',
+	});
+	const error = errors.values || {};
+	const operator = watch('operator');
+	const maxFields = getOperatorMaxFieldsAllowed(operator);
 
-	if (valuesInputsCount === 0) return null;
+	useEffect(() => {
+		if (maxFields === 0) {
+			remove();
+		} else if (!fields.length) {
+			append({ value: '' });
+		}
+	}, [maxFields]);
 
-	if (isTextType) return (<FilterFormTextValues />);
+	if (maxFields === 0) return null;
+
+	if (isTextType(type)) {
+		if (maxFields === 1) return <FormTextField name="values.0.value" formError={!!error?.[0]} />;
+		return (
+			<>
+				{fields.map((field, i) => (
+					<ArrayFieldContainer
+						key={field.id}
+						onRemove={() => remove(i)}
+						disableRemove={fields.length === 1}
+						onAdd={() => append({ value: '' })}
+						disableAdd={i !== (fields.length - 1)}
+					>
+						<FormTextField name={`values.${i}.value`} formError={!!error?.[i]} />
+					</ArrayFieldContainer>
+				))}
+			</>
+		);
+	}
 
 	return (
 		<>
 			type not created yet: {type}
-			{Array(valuesInputsCount).fill(0).map((i, index) => (
-				<FormTextField name={`values.${index}`} />
+			{range(0, Math.min(maxFields, 3)).map((i) => (
+				<FormTextField name={`values.${i}.value`} />
 			))}
 		</>
 	);
